Add ability to handle dismissed PR reviews
This change will allow a re-approval of a pull request that rebased by `renovate[bot]`.

Re-approval requires the following conditions to be met:
- The original pull request was opened by `renovate[bot]`
- The approval being dismissed was created by `renovate-approve[bot]`
- The dismissing commit was created by `renovate[bot]`
- The pull request body does not contain the string `merge this manually` which indicates the PR can be merged automatically.

I have been unable to validate this myself. It will need to be tested before being released. I have a high level of confidence that it will work as anticipated, but you can never be 100%!